CodeAlpha_HandwrittenCharRecognition/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
tensorflow
flask
Pillow
emnist
